requirements.txt
Flask==3.0.2
PyJWT==2.8.0
flask-cors==4.0.0
